Ability to configure Cookie Domain and Expiredays
A cookie can have the "Domain" attribute set to a valid domain name,
which we call the cover domain of the cookie. If the "Domain" attribute
is not set, we say the cover domain is null.

If cover domain is null, a cookie is only applicable to its origin
domain. For example, a cookie from www.cats.com is not applicable to
cats.com, and vice versa, if cover domain is null.

If cover domain is set, a cookie is applicable to the cover domain and
all its subdomains. For example the cover domain is cats.com, therefore
the cookie is applicable to cats.com, x.cats.com, x.y.cats.com, etc.

More information:
http://bayou.io/draft/cookie.domain.html