Fontawesome works inside buttons
However, my solution is not even close to pretty. I'm short on time and not that savy with vanilla-js, so I don't really know how to traverse the `event.target` properly in order to find the element that has the `cc-btn` class. I _think_ the click events just should be attached directly to the btns instead and the use the `event.currentTarget` props instead. 

Well, hopes this is hlping someone.