fix: Buttons are marked up as button elements
Having the elements `a[role=button]` is ok, however omitting the
href attribute makes it impossible to interact with the element with a
keyboard.

We could add a href with something like `javascript:void(0)` however
this isn't semantically correct. We should instead use the button
element which doesn't require the href attribute to interact with it.

Because we're now using the button element we can drop the `role=button`
as the role is implicit.