Fix handling of cc-dismiss
Due to the nature of the className that's being matched `cc-dismiss` (lowercase), the cookie doesn't get set because the matched status (`dismiss`) is lowercase instead of the uppercase string expected (`DISMISS`). This causes the validity check to fail in the `setStatuses` function and the cookie isn't set, creating a frustrating situation for the user where even after dismissing the popup, it will keep showing on each visit.