feat: default exported module is now a UMD (Universal Module Definition) module rather than CJS
The UMD module betters supports using the package directly in a browser using tools like unpkg.com. An example is in /examples-new/simplest/simplest.html.
It still works with most bundlers without changing since the main attribute of the package.json was updated to point to the new path (`"main": "dist/umd/index.js"`).
More on UMD modules at https://github.com/umdjs/umd

BREAKING CHANGE: Since this changes the path to the module and the module format is technically changing, this is marked as a breaking change. In all likelihood if you're using a bundler (webpack, rollup) this shouldn't be noticed.